cast bool to float

+++ src/gluonts/torch/distributions/implicit_quantile_network.py
@@ -131,7 +131,9 @@ class ImplicitQuantileNetwork(Distribution):
     def quantile_loss(self, value) -> torch.Tensor:
         # penalize by tau for over-predicting
         # and by 1-tau for under-predicting
-        return (self.taus - (self.outputs < value)) * (self.outputs - value)
+        return (self.taus - (self.outputs < value).float()) * (
+            self.outputs - value
+        )
 
 
 class ImplicitQuantileNetworkOutput(DistributionOutput):
